requirements.txt
requests>=2.31.0
Pillow>=10.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0 